# docs/swagger.yaml
basePath: /
definitions:
  controllers.HTTPError:
    type: object
  controllers.InputMessage:
    properties:
      messageContent:
        type: string
      userID:
        type: string
    type: object
  controllers.Intent:
    properties:
      IntentName:
        type: string
      newName:
        type: string
      prompts:
        $ref: '#/definitions/models.Prompt'
      reply:
        $ref: '#/definitions/models.ResponseMessage'
      trainingPhrases:
        items:
          type: string
        type: array
    type: object
  controllers.ResponseMessage:
    properties:
      message_content:
        type: string
      user_id:
        type: string
    type: object
  models.Prompt:
    properties:
      promptQuestion:
        items:
          type: string
        type: array
    type: object
  models.ResponseMessage:
    properties:
      message_content:
        type: string
      user_id:
        type: string
    type: object
  models.User:
    properties:
      password:
        minLength: 5
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
host: localhost:3000
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: Chatbot API with Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Chatbot API
  version: "1.0"
paths:
  /intents/ReplyIntent:
    post:
      consumes:
      - application/json
      description: Reply to an intent that is POST request from user
      parameters:
      - description: user id
        in: body
        name: inputMessage
        required: true
        schema:
          $ref: '#/definitions/controllers.InputMessage'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.HTTPError'
      summary: Reply to an intent
      tags:
      - Intents
  /intents/addIntent:
    post:
      consumes:
      - application/json
      description: Add an intent to DB
      parameters:
      - description: Name of new intent
        in: body
        name: newIntent
        required: true
        schema:
          $ref: '#/definitions/controllers.Intent'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.HTTPError'
      summary: Add an intent to DB
      tags:
      - Intents
  /intents/deleteIntent:
    delete:
      consumes:
      - application/json
      description: Delete an intent from DB, ===ONLY NEED  TO PASS IN IntentName===
      parameters:
      - description: Name of the intent that you want to delete from db
        in: body
        name: intentName
        required: true
        schema:
          $ref: '#/definitions/controllers.Intent'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.HTTPError'
      summary: Delte an intent by querying intent name
      tags:
      - Intents
  /intents/listIntent:
    post:
      description: List all intents
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.Intent'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.HTTPError'
      summary: List all intents and training phrases
      tags:
      - Intents
  /intents/modifyIntent:
    patch:
      consumes:
      - application/json
      description: Modify an intent to the body's intent by querying "IntentName".
      parameters:
      - description: The new intent, pass in NewName to change the current name
        in: body
        name: intent
        required: true
        schema:
          $ref: '#/definitions/controllers.Intent'
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.Intent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.HTTPError'
      summary: Modify an intent
      tags:
      - Intents
  /users/CheckToken:
    post:
      consumes:
      - application/json
      description: CheckToken
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.HTTPError'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/controllers.HTTPError'
      summary: CheckToken
      tags:
      - users
  /users/Login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: user1
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.HTTPError'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/controllers.HTTPError'
      summary: Login
      tags:
      - users
  /users/addUser:
    post:
      consumes:
      - application/json
      description: Modify an user.
      parameters:
      - description: user1
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.HTTPError'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/controllers.HTTPError'
      summary: Add an user
      tags:
      - users
swagger: "2.0"